Update JSUI.html
Remove nullish coalescing for wider support. I initially thought this was needed but during the same development spree I found OR to work for a following condition that was hard to figure out. It ends up working here too.

Changed function names to make sense rather than for faster typing.